requirements.txt
# AutoDownloader - Dependências Python

# Web Scraping e Automação
selenium>=4.15.0
webdriver-manager>=4.0.1

# HTTP Requests
requests>=2.31.0

# Configuração
python-dotenv>=1.0.0

# Utilitários
python-dateutil>=2.8.2

# Opcional: Testes
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Opcional: Linting e Formatação
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.0

# Opcional: Documentação
sphinx>=7.2.0
sphinx-rtd-theme>=2.0.0
